% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expfactory.rsvp.R
\name{process_rsvp}
\alias{process_rsvp}
\title{Process RSVP Task data file}
\usage{
process_rsvp(path, p)
}
\arguments{
\item{path}{Path to data file}

\item{p}{Participant identifier}
}
\value{
Data frame
}
\description{
\code{process_rsvp()} processes data files generated by the RSVP task
(\url{https://earcanal.github.io/rsvp-task/}).
}
\examples{
\dontrun{
rsvp <- expand.grid(token = participants$token) \%>\%
 rowwise() \%>\%
 mutate(path=paste0(data_dir, '/', token, '_finished/rsvp-task-results.json')) \%>\%
 do(., process_rsvp(.$path, .$token))
}
}
\keyword{attention}
\keyword{blink}
\keyword{expfactory}
\keyword{rsvp}
