% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expfactory.rsvp.R
\name{process_rsvp}
\alias{process_rsvp}
\title{Process RSVP Task data file}
\usage{
process_rsvp(path, p)
}
\arguments{
\item{path}{Path to data file}

\item{p}{Participant identifier}
}
\value{
Data frame
}
\description{
\code{process_rsvp()} processes data files generated by the RSVP task
(\url{https://earcanal.github.io/rsvp-task/}).
}
\examples{
\dontrun{
rsvp_df <- expand.grid(token = participants$token) \%>\%
 mutate(path=paste('data/', token, '_finished/rsvp-task-results.json', sep=''))
rsvp_df <- rsvp_df \%>\%
group_by(token) \%>\%
 do(., process_rsvp(.$path, .$token)) \%>\%
 ungroup \%>\%
 left_join(participants, by='token') \%>\%
 filter(phase == 'test', test_part == "response") \%>\%
 select(p, condition, trial_number, lag, correct, t1_correct, t2_correct, rt)
}
}
\keyword{attention}
\keyword{blink}
\keyword{expfactory}
\keyword{rsvp}
